Switch to using DDL for root password update
Using DDL for the root password update triggers the cluster to
replicate the password rather than relying on the charm to apply it
to each unit. This means that only the leader should update the
password.

To avoid a password update via config being missed the implicit save
of config data is turned off on update-status hooks.

Change-Id: I00fe3c7c32a000c6b697263a9ad32273515952a7
Closes-Bug: #1838124
Closes-Bug: #1838125