Ensure PROCESS perm for sstuser on series-upgrade
This patch ensures that the PROCESS permission is added to the sstuser
user during a series upgrade from xenial to bionic.  Also patches the
README to correct some of the commands in terms of order of parameters.

Closes-Bug: #1838044
(cherry picked from commit 15e5042854f3dabd59abe8913ffe06e4a72de9e7)

Change-Id: Iede2faaa1a06228e12c0b03882ddccd56bcb99f8