React to `source` or `key` changes
After changing the `source` or `key` in the configuration,
the `config-changed` hook was activated, but did not call `add_source`
and `apt_update`.
This logic has been added to the `config-changed` hook as well as
the state changed to `blocked` if the addition of a new source
or key fails.

Closes-Bug: #1893034
Change-Id: I0e7afd5a06c4945341329dd96e8eef3016367386